[FE-14] Create Pie Graph for transaction list modal
